Add camera line tracing and calulating error for the PID.
Added Camera, PID, System etc structs and enable the camera to track the
black line and calculate error by taking a line of pixels in the centre
of the screen, determining if they are black or white, then averaging
their positions. In the PID we subtract the position of the line from
the horizontal centre of the screen to yield an error value for which we
can start to correct.

+++ src/bcm2835.rs
@@ -60,9 +60,9 @@ pub fn save_picture(filename: &[i8; 5]) -> i32 {
     }
 }
 
-pub fn get_pixel(row: i32, col: i32, color: i32) -> i8 {
+pub fn get_pixel(row: u32, col: u32, color: u32) -> i8 {
     unsafe {
-        _Z9get_pixeliii(row, col, color)
+        _Z9get_pixeliii(row as i32, col as i32, color as i32)
     }
 }
